% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ylhelperfunctions.R
\name{coll_formats}
\alias{coll_formats}
\title{Apply format to dataframe columns}
\usage{
coll_formats(wb, sheet, dataframe, style, cols)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet name or number}

\item{dataframe}{dataframe to write}

\item{style}{Excel style definition}

\item{cols}{Columns to apply format}
}
\description{
Apply format to dataframe columns starting from second row up to end of dataframe
}
\examples{
coll_formats(wb, sheet, dataframe, style, cols)
}
\keyword{XLSX}
\keyword{column}
\keyword{format}
