% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ylhelperfunctions.R
\name{coll_width}
\alias{coll_width}
\title{Set width of columns}
\usage{
coll_width(wb, sheet, cols, width)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet name or number}

\item{cols}{Columns to apply format}

\item{width}{Width of column in points}
}
\description{
Set width of columns
}
\examples{
coll_width(wb, sheet, cols, width)
}
\keyword{XLSX}
\keyword{column}
\keyword{width}
