% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ylhelperfunctions.R
\name{Write_XLSX_auto_width}
\alias{Write_XLSX_auto_width}
\title{Write Dataframe to XLSX with width set at "auto"}
\usage{
Write_XLSX_auto_width(wb, sheet, dataframe)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet name or number}

\item{dataframe}{dataframe to write}
}
\description{
Write Dataframe to XLSX with column width set to Auto
}
\examples{
Write_XLSX_auto_width(wb, sheet, dataframe)
}
\keyword{XLSX}
\keyword{auto}
\keyword{columns}
\keyword{width}
