% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ylhelperfunctions.R
\name{Write_XLSX_coll_format}
\alias{Write_XLSX_coll_format}
\title{Write Dataframe to XLSX with specified format to specified columns}
\usage{
Write_XLSX_coll_format(wb, sheet, dataframe, style, cols)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet name or number}

\item{dataframe}{dataframe to write}

\item{style}{Excel style definition}

\item{cols}{Columns to apply format}
}
\description{
Write Dataframe to XLSX with specified format to specified columns
}
\examples{
Write_XLSX_coll_format(wb, sheet, dataframe, style, cols)
}
\keyword{XLSX}
\keyword{columns}
\keyword{dataframe}
\keyword{format}
